feat(Member.kt, TokenRequest.kt, ProductController.kt, GlobalExceptionHandler.kt): add validation to data objects and add a global exception handler for all controllers
- Member.kt with jakarta validation
- TokenRequest.kt with jakarta validation
- GlobalExceptionHandler.kt
- extract exceptionhandlers from ProductController.kt to GlobalExceptionHandler.kt
