+++ src/main/kotlin/ecommerce/advice/GlobalExceptionHandler.kt
@@ -1,0 +1,21 @@
+package ecommerce.advice
+
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.MethodArgumentNotValidException
+import org.springframework.web.bind.annotation.ControllerAdvice
+import org.springframework.web.bind.annotation.ExceptionHandler
+
+@ControllerAdvice
+class GlobalExceptionHandler {
+    @ExceptionHandler(MethodArgumentNotValidException::class)
+    fun handleConstraintViolations(e: MethodArgumentNotValidException): ResponseEntity<String> {
+        val errors = e.bindingResult.fieldErrors.joinToString("; ") { "${it.field}: ${it.defaultMessage}" }
+        return ResponseEntity.badRequest().body(errors)
+    }
+
+    // TODO maybe split function into handling sameNameException and other exceptions to be less generic
+    @ExceptionHandler(IllegalArgumentException::class)
+    fun handleIllegalArgument(e: IllegalArgumentException): ResponseEntity<String> {
+        return ResponseEntity.badRequest().body(e.message)
+    }
+}
